moved unsafe usage to ring/ring_vector_ops.go
