+++ bfv/encoder.go
@@ -3,8 +3,6 @@ package bfv
 
 import (
 	"fmt"
-	"math/big"
-	"unsafe"
 
 	"github.com/tuneinsight/lattigo/v3/ring"
 	"github.com/tuneinsight/lattigo/v3/utils"
@@ -65,7 +63,7 @@ type encoder struct {
 	indexMatrix []uint64
 	scaler      ring.Scaler
 
-	rescaleParams []uint64
+	tInvModQ []uint64
 
 	tmpPoly *ring.Poly
 	tmpPtRt *PlaintextRingT
@@ -107,12 +105,12 @@ func NewEncoder(params Parameters) Encoder {
 	}
 
 	return &encoder{
-		params:        params,
-		indexMatrix:   indexMatrix,
-		scaler:        ring.NewRNSScaler(ringQ, ringT),
-		rescaleParams: rescaleParams,
-		tmpPoly:       ringT.NewPoly(),
-		tmpPtRt:       NewPlaintextRingT(params),
+		params:      params,
+		indexMatrix: indexMatrix,
+		scaler:      ring.NewRNSScaler(ringQ, ringT),
+		tInvModQ:    rescaleParams,
+		tmpPoly:     ringT.NewPoly(),
+		tmpPtRt:     NewPlaintextRingT(params),
 	}
 }
 
@@ -212,7 +210,7 @@ func (ecd *encoder) EncodeIntMul(coeffs []int64, p *PlaintextMul) {
 
 // ScaleUp transforms a PlaintextRingT (R_t) into a Plaintext (R_q) by scaling up the coefficient by Q/t.
 func (ecd *encoder) ScaleUp(ptRt *PlaintextRingT, pt *Plaintext) {
-	ecd.scaleUp(ecd.params.RingQ(), ecd.params.RingT(), ecd.tmpPoly.Coeffs[0], ptRt.Value, pt.Value)
+	ring.ScaleUpVec(ecd.params.RingQ(), ecd.params.RingT(), ecd.tInvModQ, ecd.tmpPoly.Coeffs[0], ptRt.Value, pt.Value)
 }
 
 // ScaleDown transforms a Plaintext (R_q) into a PlaintextRingT (R_t) by scaling down the coefficient by t/Q and rounding.
@@ -314,63 +312,11 @@ func (ecd *encoder) DecodeIntNew(p interface{}) (coeffs []int64) {
 // Encoder can be used concurrently.
 func (ecd *encoder) ShallowCopy() Encoder {
 	return &encoder{
-		params:        ecd.params,
-		indexMatrix:   ecd.indexMatrix,
-		scaler:        ring.NewRNSScaler(ecd.params.RingQ(), ecd.params.RingT()),
-		rescaleParams: ecd.rescaleParams,
-		tmpPoly:       ecd.params.RingT().NewPoly(),
-		tmpPtRt:       NewPlaintextRingT(ecd.params),
-	}
-}
-
-// scaleUp takes m mod T and returns round((m*Q)/T) mod Q
-func (ecd *encoder) scaleUp(ringQ, ringT *ring.Ring, tmp []uint64, pIn, pOut *ring.Poly) {
-
-	qModTmontgomery := ring.MForm(new(big.Int).Mod(ringQ.ModulusBigint, ringT.ModulusBigint).Uint64(), ringT.Modulus[0], ringT.BredParams[0])
-
-	t := ringT.Modulus[0]
-	tHalf := t >> 1
-	tInv := ringT.MredParams[0]
-
-	// (x * Q + T/2) mod T
-	for i := 0; i < ringQ.N; i = i + 8 {
-		x := (*[8]uint64)(unsafe.Pointer(&pIn.Coeffs[0][i]))
-		z := (*[8]uint64)(unsafe.Pointer(&tmp[i]))
-
-		z[0] = ring.CRed(ring.MRed(x[0], qModTmontgomery, t, tInv)+tHalf, t)
-		z[1] = ring.CRed(ring.MRed(x[1], qModTmontgomery, t, tInv)+tHalf, t)
-		z[2] = ring.CRed(ring.MRed(x[2], qModTmontgomery, t, tInv)+tHalf, t)
-		z[3] = ring.CRed(ring.MRed(x[3], qModTmontgomery, t, tInv)+tHalf, t)
-		z[4] = ring.CRed(ring.MRed(x[4], qModTmontgomery, t, tInv)+tHalf, t)
-		z[5] = ring.CRed(ring.MRed(x[5], qModTmontgomery, t, tInv)+tHalf, t)
-		z[6] = ring.CRed(ring.MRed(x[6], qModTmontgomery, t, tInv)+tHalf, t)
-		z[7] = ring.CRed(ring.MRed(x[7], qModTmontgomery, t, tInv)+tHalf, t)
-	}
-
-	// (x * T^-1 - T/2) mod Qi
-	for i := 0; i < len(pOut.Coeffs); i++ {
-		p0tmp := tmp
-		p1tmp := pOut.Coeffs[i]
-		qi := ringQ.Modulus[i]
-		bredParams := ringQ.BredParams[i]
-		mredParams := ringQ.MredParams[i]
-		rescaleParams := qi - ecd.rescaleParams[i]
-
-		tHalfNegQi := qi - ring.BRedAdd(tHalf, qi, bredParams)
-
-		for j := 0; j < ringQ.N; j = j + 8 {
-
-			x := (*[8]uint64)(unsafe.Pointer(&p0tmp[j]))
-			z := (*[8]uint64)(unsafe.Pointer(&p1tmp[j]))
-
-			z[0] = ring.MRed(x[0]+tHalfNegQi, rescaleParams, qi, mredParams)
-			z[1] = ring.MRed(x[1]+tHalfNegQi, rescaleParams, qi, mredParams)
-			z[2] = ring.MRed(x[2]+tHalfNegQi, rescaleParams, qi, mredParams)
-			z[3] = ring.MRed(x[3]+tHalfNegQi, rescaleParams, qi, mredParams)
-			z[4] = ring.MRed(x[4]+tHalfNegQi, rescaleParams, qi, mredParams)
-			z[5] = ring.MRed(x[5]+tHalfNegQi, rescaleParams, qi, mredParams)
-			z[6] = ring.MRed(x[6]+tHalfNegQi, rescaleParams, qi, mredParams)
-			z[7] = ring.MRed(x[7]+tHalfNegQi, rescaleParams, qi, mredParams)
-		}
+		params:      ecd.params,
+		indexMatrix: ecd.indexMatrix,
+		scaler:      ring.NewRNSScaler(ecd.params.RingQ(), ecd.params.RingT()),
+		tInvModQ:    ecd.tInvModQ,
+		tmpPoly:     ecd.params.RingT().NewPoly(),
+		tmpPtRt:     NewPlaintextRingT(ecd.params),
 	}
 }
