+++ ckks/advanced/homomorphic_encoding.go
@@ -4,6 +4,7 @@ import (
 	"math"
 
 	"github.com/tuneinsight/lattigo/v3/ckks"
+	"github.com/tuneinsight/lattigo/v3/ring"
 	"github.com/tuneinsight/lattigo/v3/utils"
 )
 
@@ -499,14 +500,14 @@ func genFFTDiagMatrix(logL, fftLevel int, a, b, c []complex128, ltType LinearTra
 	vectors = make(map[int][]complex128)
 
 	if bitreversed {
-		ckks.SliceBitReverseInPlaceComplex128(a, 1<<logL)
-		ckks.SliceBitReverseInPlaceComplex128(b, 1<<logL)
-		ckks.SliceBitReverseInPlaceComplex128(c, 1<<logL)
+		ring.SliceBitReverseInPlaceComplex128(a, 1<<logL)
+		ring.SliceBitReverseInPlaceComplex128(b, 1<<logL)
+		ring.SliceBitReverseInPlaceComplex128(c, 1<<logL)
 
 		if len(a) > 1<<logL {
-			ckks.SliceBitReverseInPlaceComplex128(a[1<<logL:], 1<<logL)
-			ckks.SliceBitReverseInPlaceComplex128(b[1<<logL:], 1<<logL)
-			ckks.SliceBitReverseInPlaceComplex128(c[1<<logL:], 1<<logL)
+			ring.SliceBitReverseInPlaceComplex128(a[1<<logL:], 1<<logL)
+			ring.SliceBitReverseInPlaceComplex128(b[1<<logL:], 1<<logL)
+			ring.SliceBitReverseInPlaceComplex128(c[1<<logL:], 1<<logL)
 		}
 	}
 
@@ -551,14 +552,14 @@ func multiplyFFTMatrixWithNextFFTLevel(vec map[int][]complex128, logL, N, nextLe
 	}
 
 	if bitreversed {
-		ckks.SliceBitReverseInPlaceComplex128(a, 1<<logL)
-		ckks.SliceBitReverseInPlaceComplex128(b, 1<<logL)
-		ckks.SliceBitReverseInPlaceComplex128(c, 1<<logL)
+		ring.SliceBitReverseInPlaceComplex128(a, 1<<logL)
+		ring.SliceBitReverseInPlaceComplex128(b, 1<<logL)
+		ring.SliceBitReverseInPlaceComplex128(c, 1<<logL)
 
 		if len(a) > 1<<logL {
-			ckks.SliceBitReverseInPlaceComplex128(a[1<<logL:], 1<<logL)
-			ckks.SliceBitReverseInPlaceComplex128(b[1<<logL:], 1<<logL)
-			ckks.SliceBitReverseInPlaceComplex128(c[1<<logL:], 1<<logL)
+			ring.SliceBitReverseInPlaceComplex128(a[1<<logL:], 1<<logL)
+			ring.SliceBitReverseInPlaceComplex128(b[1<<logL:], 1<<logL)
+			ring.SliceBitReverseInPlaceComplex128(c[1<<logL:], 1<<logL)
 		}
 	}
