+++ ckks/utils.go
@@ -290,69 +290,3 @@ func GenSwitchkeysRescalingParams(Q, P []uint64) (params []uint64) {
 
 	return
 }
-
-// SliceBitReverseInPlaceFloat64 applies an in-place bit-reverse permuation on the input slice.
-func SliceBitReverseInPlaceFloat64(slice []float64, N int) {
-
-	var bit, j int
-
-	for i := 1; i < N; i++ {
-
-		bit = N >> 1
-
-		for j >= bit {
-			j -= bit
-			bit >>= 1
-		}
-
-		j += bit
-
-		if i < j {
-			slice[i], slice[j] = slice[j], slice[i]
-		}
-	}
-}
-
-// SliceBitReverseInPlaceComplex128 applies an in-place bit-reverse permuation on the input slice.
-func SliceBitReverseInPlaceComplex128(slice []complex128, N int) {
-
-	var bit, j int
-
-	for i := 1; i < N; i++ {
-
-		bit = N >> 1
-
-		for j >= bit {
-			j -= bit
-			bit >>= 1
-		}
-
-		j += bit
-
-		if i < j {
-			slice[i], slice[j] = slice[j], slice[i]
-		}
-	}
-}
-
-// SliceBitReverseInPlaceRingComplex applies an in-place bit-reverse permuation on the input slice.
-func SliceBitReverseInPlaceRingComplex(slice []*ring.Complex, N int) {
-
-	var bit, j int
-
-	for i := 1; i < N; i++ {
-
-		bit = N >> 1
-
-		for j >= bit {
-			j -= bit
-			bit >>= 1
-		}
-
-		j += bit
-
-		if i < j {
-			slice[i], slice[j] = slice[j], slice[i]
-		}
-	}
-}
