+++ ring/ring_vector_ops.go
@@ -1,6 +1,8 @@
 package ring
 
 import (
+	"math/big"
+	"math/bits"
 	"unsafe"
 )
 
@@ -618,5 +620,360 @@ func MulByPow2Vec(p1, p2 []uint64, pow2 int, qi, mredParams uint64) {
 		z[5] = PowerOf2(x[5], pow2, qi, mredParams)
 		z[6] = PowerOf2(x[6], pow2, qi, mredParams)
 		z[7] = PowerOf2(x[7], pow2, qi, mredParams)
+	}
+}
+
+// ScaleUpVec takes a Poly pIn in ringT, scales its coefficients up by (Q/T) mod Q, and writes the result in a
+// Poly pOut in ringQ.
+func ScaleUpVec(ringQ, ringT *Ring, rescaleParams, tmp []uint64, pIn, pOut *Poly) {
+
+	qModTmontgomery := MForm(new(big.Int).Mod(ringQ.ModulusBigint, ringT.ModulusBigint).Uint64(), ringT.Modulus[0], ringT.BredParams[0])
+
+	t := ringT.Modulus[0]
+	tHalf := t >> 1
+	tInv := ringT.MredParams[0]
+
+	// (x * Q + T/2) mod T
+	for i := 0; i < ringQ.N; i = i + 8 {
+		x := (*[8]uint64)(unsafe.Pointer(&pIn.Coeffs[0][i]))
+		z := (*[8]uint64)(unsafe.Pointer(&tmp[i]))
+
+		z[0] = CRed(MRed(x[0], qModTmontgomery, t, tInv)+tHalf, t)
+		z[1] = CRed(MRed(x[1], qModTmontgomery, t, tInv)+tHalf, t)
+		z[2] = CRed(MRed(x[2], qModTmontgomery, t, tInv)+tHalf, t)
+		z[3] = CRed(MRed(x[3], qModTmontgomery, t, tInv)+tHalf, t)
+		z[4] = CRed(MRed(x[4], qModTmontgomery, t, tInv)+tHalf, t)
+		z[5] = CRed(MRed(x[5], qModTmontgomery, t, tInv)+tHalf, t)
+		z[6] = CRed(MRed(x[6], qModTmontgomery, t, tInv)+tHalf, t)
+		z[7] = CRed(MRed(x[7], qModTmontgomery, t, tInv)+tHalf, t)
+	}
+
+	// (x * T^-1 - T/2) mod Qi
+	for i := 0; i < len(pOut.Coeffs); i++ {
+		p0tmp := tmp
+		p1tmp := pOut.Coeffs[i]
+		qi := ringQ.Modulus[i]
+		bredParams := ringQ.BredParams[i]
+		mredParams := ringQ.MredParams[i]
+		rescaleParams := qi - rescaleParams[i]
+
+		tHalfNegQi := qi - BRedAdd(tHalf, qi, bredParams)
+
+		for j := 0; j < ringQ.N; j = j + 8 {
+
+			x := (*[8]uint64)(unsafe.Pointer(&p0tmp[j]))
+			z := (*[8]uint64)(unsafe.Pointer(&p1tmp[j]))
+
+			z[0] = MRed(x[0]+tHalfNegQi, rescaleParams, qi, mredParams)
+			z[1] = MRed(x[1]+tHalfNegQi, rescaleParams, qi, mredParams)
+			z[2] = MRed(x[2]+tHalfNegQi, rescaleParams, qi, mredParams)
+			z[3] = MRed(x[3]+tHalfNegQi, rescaleParams, qi, mredParams)
+			z[4] = MRed(x[4]+tHalfNegQi, rescaleParams, qi, mredParams)
+			z[5] = MRed(x[5]+tHalfNegQi, rescaleParams, qi, mredParams)
+			z[6] = MRed(x[6]+tHalfNegQi, rescaleParams, qi, mredParams)
+			z[7] = MRed(x[7]+tHalfNegQi, rescaleParams, qi, mredParams)
+		}
+	}
+}
+
+// SpecialFFTVec performs the CKKS special FFT transform in place.
+func SpecialFFTVec(values []complex128, N, M int, rotGroup []int, roots []complex128) {
+
+	SliceBitReverseInPlaceComplex128(values, N)
+
+	logN := int(bits.Len64(uint64(N))) - 1
+	logM := int(bits.Len64(uint64(M))) - 1
+
+	for loglen := 1; loglen <= logN; loglen++ {
+		len := 1 << loglen
+		lenh := len >> 1
+		lenq := len << 2
+		logGap := logM - 2 - loglen
+		mask := lenq - 1
+
+		if lenh > 8 {
+			for i := 0; i < N; i += len {
+
+				for j, k := 0, i; j < lenh; j, k = j+8, k+8 {
+
+					u := (*[8]complex128)(unsafe.Pointer(&values[k]))
+					v := (*[8]complex128)(unsafe.Pointer(&values[k+lenh]))
+					w := (*[8]int)(unsafe.Pointer(&rotGroup[j]))
+
+					v[0] *= roots[(w[0]&mask)<<logGap]
+					v[1] *= roots[(w[1]&mask)<<logGap]
+					v[2] *= roots[(w[2]&mask)<<logGap]
+					v[3] *= roots[(w[3]&mask)<<logGap]
+					v[4] *= roots[(w[4]&mask)<<logGap]
+					v[5] *= roots[(w[5]&mask)<<logGap]
+					v[6] *= roots[(w[6]&mask)<<logGap]
+					v[7] *= roots[(w[7]&mask)<<logGap]
+
+					u[0], v[0] = u[0]+v[0], u[0]-v[0]
+					u[1], v[1] = u[1]+v[1], u[1]-v[1]
+					u[2], v[2] = u[2]+v[2], u[2]-v[2]
+					u[3], v[3] = u[3]+v[3], u[3]-v[3]
+					u[4], v[4] = u[4]+v[4], u[4]-v[4]
+					u[5], v[5] = u[5]+v[5], u[5]-v[5]
+					u[6], v[6] = u[6]+v[6], u[6]-v[6]
+					u[7], v[7] = u[7]+v[7], u[7]-v[7]
+				}
+			}
+		} else if lenh == 8 {
+
+			psi0 := roots[(rotGroup[0]&mask)<<logGap]
+			psi1 := roots[(rotGroup[1]&mask)<<logGap]
+			psi2 := roots[(rotGroup[2]&mask)<<logGap]
+			psi3 := roots[(rotGroup[3]&mask)<<logGap]
+			psi4 := roots[(rotGroup[4]&mask)<<logGap]
+			psi5 := roots[(rotGroup[5]&mask)<<logGap]
+			psi6 := roots[(rotGroup[6]&mask)<<logGap]
+			psi7 := roots[(rotGroup[7]&mask)<<logGap]
+
+			for i := 0; i < N; i += 16 {
+
+				u := (*[16]complex128)(unsafe.Pointer(&values[i]))
+
+				u[8] *= psi0
+				u[9] *= psi1
+				u[10] *= psi2
+				u[11] *= psi3
+				u[12] *= psi4
+				u[13] *= psi5
+				u[14] *= psi6
+				u[15] *= psi7
+
+				u[0], u[8] = u[0]+u[8], u[0]-u[8]
+				u[1], u[9] = u[1]+u[9], u[1]-u[9]
+				u[2], u[10] = u[2]+u[10], u[2]-u[10]
+				u[3], u[11] = u[3]+u[11], u[3]-u[11]
+				u[4], u[12] = u[4]+u[12], u[4]-u[12]
+				u[5], u[13] = u[5]+u[13], u[5]-u[13]
+				u[6], u[14] = u[6]+u[14], u[6]-u[14]
+				u[7], u[15] = u[7]+u[15], u[7]-u[15]
+
+			}
+		} else if lenh == 4 {
+
+			psi0 := roots[(rotGroup[0]&mask)<<logGap]
+			psi1 := roots[(rotGroup[1]&mask)<<logGap]
+			psi2 := roots[(rotGroup[2]&mask)<<logGap]
+			psi3 := roots[(rotGroup[3]&mask)<<logGap]
+
+			for i := 0; i < N; i += 16 {
+
+				u := (*[16]complex128)(unsafe.Pointer(&values[i]))
+
+				u[4] *= psi0
+				u[5] *= psi1
+				u[6] *= psi2
+				u[7] *= psi3
+				u[12] *= psi0
+				u[13] *= psi1
+				u[14] *= psi2
+				u[15] *= psi3
+
+				u[0], u[4] = u[0]+u[4], u[0]-u[4]
+				u[1], u[5] = u[1]+u[5], u[1]-u[5]
+				u[2], u[6] = u[2]+u[6], u[2]-u[6]
+				u[3], u[7] = u[3]+u[7], u[3]-u[7]
+				u[8], u[12] = u[8]+u[12], u[8]-u[12]
+				u[9], u[13] = u[9]+u[13], u[9]-u[13]
+				u[10], u[14] = u[10]+u[14], u[10]-u[14]
+				u[11], u[15] = u[11]+u[15], u[11]-u[15]
+			}
+		} else if lenh == 2 {
+
+			psi0 := roots[(rotGroup[0]&mask)<<logGap]
+			psi1 := roots[(rotGroup[1]&mask)<<logGap]
+
+			for i := 0; i < N; i += 16 {
+
+				u := (*[16]complex128)(unsafe.Pointer(&values[i]))
+
+				u[2] *= psi0
+				u[3] *= psi1
+				u[6] *= psi0
+				u[7] *= psi1
+				u[10] *= psi0
+				u[11] *= psi1
+				u[14] *= psi0
+				u[15] *= psi1
+
+				u[0], u[2] = u[0]+u[2], u[0]-u[2]
+				u[1], u[3] = u[1]+u[3], u[1]-u[3]
+				u[4], u[6] = u[4]+u[6], u[4]-u[6]
+				u[5], u[7] = u[5]+u[7], u[5]-u[7]
+				u[8], u[10] = u[8]+u[10], u[8]-u[10]
+				u[9], u[11] = u[9]+u[11], u[9]-u[11]
+				u[12], u[14] = u[12]+u[14], u[12]-u[14]
+				u[13], u[15] = u[13]+u[15], u[13]-u[15]
+			}
+		} else if lenh == 1 {
+
+			psi0 := roots[(rotGroup[0]&mask)<<logGap]
+
+			for i := 0; i < N; i += 16 {
+
+				u := (*[16]complex128)(unsafe.Pointer(&values[i]))
+
+				u[1] *= psi0
+				u[3] *= psi0
+				u[5] *= psi0
+				u[7] *= psi0
+				u[9] *= psi0
+				u[11] *= psi0
+				u[13] *= psi0
+				u[15] *= psi0
+
+				u[0], u[1] = u[0]+u[1], u[0]-u[1]
+				u[2], u[3] = u[2]+u[3], u[2]-u[3]
+				u[4], u[5] = u[4]+u[5], u[4]-u[5]
+				u[6], u[7] = u[6]+u[7], u[6]-u[7]
+				u[8], u[9] = u[8]+u[9], u[8]-u[9]
+				u[10], u[11] = u[10]+u[11], u[10]-u[11]
+				u[12], u[13] = u[12]+u[13], u[12]-u[13]
+				u[14], u[15] = u[14]+u[15], u[14]-u[15]
+			}
+		}
+	}
+}
+
+// SpecialInvFFTVec performs the CKKS special inverse FFT transform in place.
+func SpecialInvFFTVec(values []complex128, N, M int, rotGroup []int, roots []complex128) {
+
+	var lenh, lenq, mask, logGap int
+
+	logN := int(bits.Len64(uint64(N))) - 1
+	logM := int(bits.Len64(uint64(M))) - 1
+
+	for loglen := logN; loglen > 0; loglen-- {
+		len := 1 << loglen
+		lenh = len >> 1
+		lenq = len << 2
+		logGap = logM - 2 - loglen
+		mask = lenq - 1
+
+		if lenh > 8 {
+			for i := 0; i < N; i += len {
+				for j, k := 0, i; j < lenh; j, k = j+8, k+8 {
+
+					u := (*[8]complex128)(unsafe.Pointer(&values[k]))
+					v := (*[8]complex128)(unsafe.Pointer(&values[k+lenh]))
+					w := (*[8]int)(unsafe.Pointer(&rotGroup[j]))
+
+					u[0], v[0] = u[0]+v[0], (u[0]-v[0])*roots[(lenq-(w[0]&mask))<<logGap]
+					u[1], v[1] = u[1]+v[1], (u[1]-v[1])*roots[(lenq-(w[1]&mask))<<logGap]
+					u[2], v[2] = u[2]+v[2], (u[2]-v[2])*roots[(lenq-(w[2]&mask))<<logGap]
+					u[3], v[3] = u[3]+v[3], (u[3]-v[3])*roots[(lenq-(w[3]&mask))<<logGap]
+					u[4], v[4] = u[4]+v[4], (u[4]-v[4])*roots[(lenq-(w[4]&mask))<<logGap]
+					u[5], v[5] = u[5]+v[5], (u[5]-v[5])*roots[(lenq-(w[5]&mask))<<logGap]
+					u[6], v[6] = u[6]+v[6], (u[6]-v[6])*roots[(lenq-(w[6]&mask))<<logGap]
+					u[7], v[7] = u[7]+v[7], (u[7]-v[7])*roots[(lenq-(w[7]&mask))<<logGap]
+				}
+			}
+		} else if lenh == 8 {
+
+			psi0 := roots[(lenq-(rotGroup[0]&mask))<<logGap]
+			psi1 := roots[(lenq-(rotGroup[1]&mask))<<logGap]
+			psi2 := roots[(lenq-(rotGroup[2]&mask))<<logGap]
+			psi3 := roots[(lenq-(rotGroup[3]&mask))<<logGap]
+			psi4 := roots[(lenq-(rotGroup[4]&mask))<<logGap]
+			psi5 := roots[(lenq-(rotGroup[5]&mask))<<logGap]
+			psi6 := roots[(lenq-(rotGroup[6]&mask))<<logGap]
+			psi7 := roots[(lenq-(rotGroup[7]&mask))<<logGap]
+
+			for i := 0; i < N; i += 16 {
+
+				u := (*[16]complex128)(unsafe.Pointer(&values[i]))
+
+				u[0], u[8] = u[0]+u[8], (u[0]-u[8])*psi0
+				u[1], u[9] = u[1]+u[9], (u[1]-u[9])*psi1
+				u[2], u[10] = u[2]+u[10], (u[2]-u[10])*psi2
+				u[3], u[11] = u[3]+u[11], (u[3]-u[11])*psi3
+				u[4], u[12] = u[4]+u[12], (u[4]-u[12])*psi4
+				u[5], u[13] = u[5]+u[13], (u[5]-u[13])*psi5
+				u[6], u[14] = u[6]+u[14], (u[6]-u[14])*psi6
+				u[7], u[15] = u[7]+u[15], (u[7]-u[15])*psi7
+			}
+
+		} else if lenh == 4 {
+
+			psi0 := roots[(lenq-(rotGroup[0]&mask))<<logGap]
+			psi1 := roots[(lenq-(rotGroup[1]&mask))<<logGap]
+			psi2 := roots[(lenq-(rotGroup[2]&mask))<<logGap]
+			psi3 := roots[(lenq-(rotGroup[3]&mask))<<logGap]
+
+			for i := 0; i < N; i += 16 {
+
+				u := (*[16]complex128)(unsafe.Pointer(&values[i]))
+
+				u[0], u[4] = u[0]+u[4], (u[0]-u[4])*psi0
+				u[1], u[5] = u[1]+u[5], (u[1]-u[5])*psi1
+				u[2], u[6] = u[2]+u[6], (u[2]-u[6])*psi2
+				u[3], u[7] = u[3]+u[7], (u[3]-u[7])*psi3
+				u[8], u[12] = u[8]+u[12], (u[8]-u[12])*psi0
+				u[9], u[13] = u[9]+u[13], (u[9]-u[13])*psi1
+				u[10], u[14] = u[10]+u[14], (u[10]-u[14])*psi2
+				u[11], u[15] = u[11]+u[15], (u[11]-u[15])*psi3
+			}
+		} else if lenh == 2 {
+
+			psi0 := roots[(lenq-(rotGroup[0]&mask))<<logGap]
+			psi1 := roots[(lenq-(rotGroup[1]&mask))<<logGap]
+
+			for i := 0; i < N; i += 16 {
+
+				u := (*[16]complex128)(unsafe.Pointer(&values[i]))
+
+				u[0], u[2] = u[0]+u[2], (u[0]-u[2])*psi0
+				u[1], u[3] = u[1]+u[3], (u[1]-u[3])*psi1
+				u[4], u[6] = u[4]+u[6], (u[4]-u[6])*psi0
+				u[5], u[7] = u[5]+u[7], (u[5]-u[7])*psi1
+				u[8], u[10] = u[8]+u[10], (u[8]-u[10])*psi0
+				u[9], u[11] = u[9]+u[11], (u[9]-u[11])*psi1
+				u[12], u[14] = u[12]+u[14], (u[12]-u[14])*psi0
+				u[13], u[15] = u[13]+u[15], (u[13]-u[15])*psi1
+			}
+		} else if lenh == 1 {
+
+			psi0 := roots[(lenq-(rotGroup[0]&mask))<<logGap]
+
+			for i := 0; i < N; i += 16 {
+
+				u := (*[16]complex128)(unsafe.Pointer(&values[i]))
+
+				u[0], u[1] = u[0]+u[1], (u[0]-u[1])*psi0
+				u[2], u[3] = u[2]+u[3], (u[2]-u[3])*psi0
+				u[4], u[5] = u[4]+u[5], (u[4]-u[5])*psi0
+				u[6], u[7] = u[6]+u[7], (u[6]-u[7])*psi0
+				u[8], u[9] = u[8]+u[9], (u[8]-u[9])*psi0
+				u[10], u[11] = u[10]+u[11], (u[10]-u[11])*psi0
+				u[12], u[13] = u[12]+u[13], (u[12]-u[13])*psi0
+				u[14], u[15] = u[14]+u[15], (u[14]-u[15])*psi0
+			}
+		}
+	}
+
+	DivideComplex128SliceVec(values, complex(float64(N), 0))
+
+	SliceBitReverseInPlaceComplex128(values, N)
+}
+
+// DivideComplex128SliceVec divides the entries in values by scaleVal in place.
+func DivideComplex128SliceVec(values []complex128, scaleVal complex128) {
+	lenValues := len(values)
+	for i := 0; i < lenValues; i = i + 8 {
+
+		v := (*[8]complex128)(unsafe.Pointer(&values[i]))
+
+		v[0] /= scaleVal
+		v[1] /= scaleVal
+		v[2] /= scaleVal
+		v[3] /= scaleVal
+		v[4] /= scaleVal
+		v[5] /= scaleVal
+		v[6] /= scaleVal
+		v[7] /= scaleVal
 	}
 }
